Image provider added

+++ target.cpp
@@ -9,7 +9,7 @@ ShootError Target::load(const std::string& filename) {
   } else {
     width_ = original_.cols;
     height_ = original_.rows;
-    printf("Image size: x: %zu, y: %zu\n", width_, height_);
+    printf("Image size: x: %d, y: %d\n", width_, height_);
   }
   cv::cvtColor(original_, grey_, CV_BGR2GRAY);
   return ShootError::NoError;
@@ -28,21 +28,20 @@ void Target::set_description(std::vector<uint16_t>&& target_description) {
 }
 
 ShootError Target::find_center() {
-
   // Check if we should start from black "10"
   // Assume so
   bool blackCenter = grey_.at<uchar>(height_ / 2, width_ / 2) < kBWLimit;
 
   if (blackCenter) {
     int xr{}, xl{}, yu{}, yd{};
-    for (size_t c{0}; c < width_ / 2; ++c) {
-
+    for (int c{0}; c < width_ / 2; ++c) {
       // Print center search cross
 
-      //if (!xr) cv::circle(original_, cv::Point(width_ / 2 + c, height_ / 2), 1, cv::Scalar(255, 0, 0));
-      //if (!xl) cv::circle(original_, cv::Point(width_ / 2 - c, height_ / 2), 1, cv::Scalar(255, 0, 0));
-      //if (!yd) cv::circle(original_, cv::Point(width_ / 2, height_ / 2 + c), 1, cv::Scalar(255, 0, 0));
-      //if (!yu) cv::circle(original_, cv::Point(width_ / 2, height_ / 2 - c), 1, cv::Scalar(255, 0, 0));
+      // if (!xr) cv::circle(original_, cv::Point(width_ / 2 + c, height_ / 2), 1, cv::Scalar(255,
+      // 0, 0)); if (!xl) cv::circle(original_, cv::Point(width_ / 2 - c, height_ / 2), 1,
+      // cv::Scalar(255, 0, 0)); if (!yd) cv::circle(original_, cv::Point(width_ / 2, height_ / 2 +
+      // c), 1, cv::Scalar(255, 0, 0)); if (!yu) cv::circle(original_, cv::Point(width_ / 2, height_
+      // / 2 - c), 1, cv::Scalar(255, 0, 0));
 
       // Actual search
 
@@ -80,7 +79,7 @@ ShootError Target::find_points() {
   bool is_white{false};
   size_t target_index{0};
 
-  for (size_t r{0}; r < height_ / 2; ++r) {
+  for (int r{0}; r < height_ / 2; ++r) {
     bool found{false};
     uchar value = grey_.at<uchar>(center_.y + r, center_.x + r);
 
@@ -106,7 +105,7 @@ ShootError Target::find_points() {
 
     if (found && target_description_[++target_index] < ColorChange::Start) {
       auto radius = sqrt(r * r + r * r);
-      target_points_[target_description_[target_index]] = static_cast<size_t>(radius);
+      target_points_[target_description_[target_index]] = static_cast<int>(radius);
       if (target_description_[++target_index] == ColorChange::Finish) {
         break;
       }
@@ -116,11 +115,9 @@ ShootError Target::find_points() {
 }
 
 void Target::print_points() {
-  /*
   for (const auto& [point, radius] : target_points_) {
     std::cout << point << "\t=>\t" << radius << std::endl;
   }
-  */
 }
 
 void Target::add_markers() {
@@ -131,11 +128,9 @@ void Target::add_markers() {
     auto font = cv::HersheyFonts::FONT_HERSHEY_SIMPLEX;
     int base_line;
     cv::Size text_size = cv::getTextSize(std::to_string(point), font, 1, 1, &base_line);
-    
-    cv::putText(
-        original_, std::to_string(point),
-        cv::Point(center_.x + radius - text_size.width, center_.y + text_size.height / 2), font,
-        1, cv::Scalar(0, 255, 0), 3);
-        
+
+    cv::putText(original_, std::to_string(point),
+                cv::Point(center_.x + radius - text_size.width, center_.y + text_size.height / 2),
+                font, 1, cv::Scalar(0, 255, 0), 3);
   }
 }
